+++ lib/lib/_read.js
@@ -1,0 +1,16 @@
+function read (size) {
+  let pushSize = Math.min(size, this.bufmax);
+  if (this.html.length > 0) {
+    if (this.debug) console.error('>>>_read size=' + pushSize + ', html="' + this.html.substr(0, pushSize) + '"');
+    this.push(this.html.substr(0, pushSize));
+    this.html = this.html.slice(pushSize);
+    if (this.done && this.html.length === 0) {
+      this.push(null); // eof
+    }
+  } else {
+    if (this.debug) console.error('>>>_read - nothing to push');
+  }
+  return;
+};
+
+module.exports = read;
