+++ lib/lib/cat.js
@@ -1,0 +1,5 @@
+function cat(data) {
+  this.html += data;
+}
+
+module.exports = cat;
